fix(web): move fetch image to new file
chore(web): update error message
